# meta/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Sportfest backend",
    "version": "1.0.0",
    "title": "Sportfest Backend",
    "contact": {
      "url": "https://github.com/atiw-sportfest"
    }
  },
  "host": "sportfest.atiw.de",
  "basePath": "/backend",
...
  "schemes": [
    "https"
  ],
  "parameters": {
...
    "did": {
      "in": "path",
      "name": "did",
      "description": "Disziplin-ID",
      "required": true,
      "type": "integer",
      "format": "int64"
    },
    "eid": {
      "in": "path",
      "name": "eid",
      "description": "Ergebnis-ID",
      "required": true,
      "type": "integer",
      "format": "int64"
    },
    "kid": {
      "in": "path",
      "name": "kid",
      "description": "Klassen-ID",
      "required": true,
      "type": "integer",
      "format": "int64"
    },
...
    "versusid": {
      "in": "path",
      "name": "vid",
      "description": "Versus-ID",
      "required": true,
      "type": "integer",
      "format": "int64"
    },
...
  },
  "responses": {
...
    "Anmeldungen": {
      "description": "Anmeldungen",
      "schema": {
        "type": "array",
        "items": {
          "$ref": "#/definitions/Anmeldung"
        }
      }
    },
    "Disziplin": {
      "description": "Disziplin erfolgreich erstellt",
      "schema": {
        "$ref": "#/definitions/Disziplin"
      }
    },
    "Ergebnis": {
      "description": "Ergebnis",
      "schema": {
        "$ref": "#/definitions/Ergebnis"
      }
    },
    "Ergebnisse": {
      "description": "Ergebnisse",
      "schema": {
        "type": "array",
        "items": {
          "$ref": "#/definitions/Ergebnis"
        }
      }
    },
...
    "Leistungen": {
      "description": "Leistungen",
      "schema": {
        "type": "array",
        "items": {
          "$ref": "#/definitions/Leistung"
        }
      }
    },
    "Klasse": {
      "description": "Klasse",
      "schema": {
        "$ref": "#/definitions/Klasse"
      }
    },
...
    "Schueler_pl": {
      "description": "Schueler_pl",
      "schema": {
        "type": "array",
        "items": {
          "$ref": "#/definitions/Schueler"
        }
      }
    }
  },
  "paths": {
...
    "/disziplin": {
      "get": {
        "summary": "Disziplinen auflisten",
        "tags": [
          "Disziplin"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Disziplin"
              }
            }
          }
        }
      },
      "post": {
        "summary": "Disziplin anlegen",
        "tags": [
          "Disziplin"
        ],
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "disziplin",
            "schema": {
              "$ref": "#/definitions/Disziplin"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/Disziplin"
          }
        }
      }
    },
    "/disziplin/{did}": {
      "get": {
        "summary": "Disziplin anzeigen",
        "tags": [
          "Disziplin"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "$ref": "#/parameters/did"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/Disziplin"
          }
        }
      },
      "post": {
        "summary": "Disziplin bearbeiten",
        "tags": [
          "Disziplin"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "$ref": "#/parameters/did"
          },
          {
            "in": "body",
            "name": "body",
            "description": "Neue Disziplin",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Disziplin"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/Disziplin"
          }
        }
      },
      "delete": {
        "summary": "Disziplin löschen",
        "tags": [
          "Disziplin"
        ],
        "parameters": [
          {
            "$ref": "#/parameters/did"
          }
        ],
        "responses": {
          "204": {
            "description": "Disziplin gelöscht"
          }
        }
      }
    },
    "/disziplin/{did}/ergebnisse": {
      "get": {
        "summary": "Ergebnisse einer Disziplin anzeigen",
        "tags": [
          "Disziplin",
          "Ergebnis"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "$ref": "#/parameters/did"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/Ergebnisse"
          }
        }
      },
      "post": {
        "summary": "Ergebnis-Leistungen auswerten",
        "tags": [
          "Disziplin",
          "Ergebnis"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "teilnehmer",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "teilnehmer": {
                  "type": "array",
                  "items": {
                    "type": "integer",
                    "format": "int64"
                  }
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/Ergebnisse"
          }
        }
      }
    },
    "/disziplin/{did}/anmeldungen": {
      "get": {
        "summary": "Anmeldungen einer Disziplin anzeigen",
        "tags": [
          "Disziplin",
          "Anmeldung"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "$ref": "#/parameters/did"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/Anmeldungen"
          }
        }
      }
    },
    "/disziplin/{did}/leistungen": {
      "get": {
        "summary": "Leistungen einer Disziplin anzeigen",
        "tags": [
          "Disziplin",
          "Ergebnis"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "$ref": "#/parameters/did"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/Leistungen"
          }
        }
      }
    },
...
    "/ergebnis/{eid}": {
      "get": {
        "summary": "Ergebnis anzeigen",
        "tags": [
          "Ergebnis"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "$ref": "#/parameters/eid"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/Ergebnis"
          }
        }
      },
      "delete": {
        "summary": "Ergebnis löschen",
        "tags": [
          "Ergebnis"
        ],
        "parameters": [
          {
            "$ref": "#/parameters/eid"
          }
        ],
        "responses": {
          "204": {
            "description": "Ergebnis gelöscht"
          }
        }
      }
    },
    "/ergebnis/versus/{vid}": {
      "get": {
        "summary": "Ergebnisse eines Versus anzeigen",
        "tags": [
          "Ergebnis"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "$ref": "#/parameters/versusid"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/Ergebnisse"
          }
        }
      }
    },
...
    "/klasse/{kid}": {
      "get": {
        "summary": "Klasse abrufen",
        "tags": [
          "Teilnehmer"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "$ref": "#/parameters/kid"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/Klasse"
          }
        }
      },
      "delete": {
        "summary": "Klasse löschen",
        "tags": [
          "Teilnehmer"
        ],
        "parameters": [
          {
            "$ref": "#/parameters/kid"
          }
        ],
        "responses": {
          "204": {
            "description": "Gelöscht."
          }
        }
      }
    },
    "/klasse/{kid}/anmeldungen": {
      "get": {
        "summary": "Anmeldungen einer Klasse anzuzeigen",
        "tags": [
          "Anmeldung"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "$ref": "#/parameters/kid"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/Anmeldungen"
          }
        }
      }
    },
    "/klasse/{kid}/ergebnisse": {
      "get": {
        "summary": "Ergebnisse einer Klasse anzuzeigen",
        "tags": [
          "Teilnehmer"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "$ref": "#/parameters/kid"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/Ergebnisse"
          }
        }
      }
    },
    "/klasse/{kid}/ergebnisse/{did}": {
      "get": {
        "summary": "Ergebnisse einer Klasse fuer eine Disziplin anzeigen",
        "tags": [
          "Teilnehmer"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "$ref": "#/parameters/did"
          },
          {
            "$ref": "#/parameters/kid"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/Ergebnisse"
          }
        }
      }
    },
    "/klasse/{kid}/schueler": {
      "get": {
        "summary": "Schueler einer Klasse anzuzeigen",
        "tags": [
          "Teilnehmer"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "$ref": "#/parameters/kid"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/Schueler_pl"
          }
        }
      }
    },
...
  },
  "securityDefinitions": {
    "jwt": {
      "description": "JWT Bearer-Authentication",
      "type": "apiKey",
      "name": "HTTP-Authorization",
      "in": "header"
    }
  },
  "definitions": {
    "Anmeldung": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "disziplin": {
          "$ref": "#/definitions/Disziplin"
        },
        "schueler": {
          "$ref": "#/definitions/Schueler"
        }
      }
    },
    "Disziplin": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "bezeichnung": {
          "type": "string"
        },
        "beschreibung": {
          "type": "string"
        },
        "team": {
          "type": "boolean"
        },
        "versus": {
          "type": "boolean"
        },
        "variablen": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Variable"
          }
        },
        "regeln": {
          "type": "string"
        }
      }
    },
    "Ergebnis": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "punkte": {
          "type": "integer"
        },
        "leistungen": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Leistung"
          }
        },
        "klasse": {
          "$ref": "#/definitions/Klasse"
        },
        "schueler": {
          "$ref": "#/definitions/Schueler"
        },
        "disziplin": {
          "$ref": "#/definitions/Disziplin"
        }
      }
    },
    "Klasse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "bezeichnung": {
          "type": "string"
        }
      }
    },
    "Leistung": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "wert": {
          "type": "string"
        },
        "variable": {
          "$ref": "#/definitions/Variable"
        }
      }
    },
    "Schueler": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "vorname": {
          "type": "string"
        },
        "nachname": {
          "type": "string"
        },
        "geschlecht": {
          "type": "string"
        },
        "klasse": {
          "$ref": "#/definitions/Klasse"
        }
      }
    },
    "Typ": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "datentyp": {
          "type": "string",
          "enum": [
            "string",
            "int",
            "float",
            "long",
            "double"
          ]
        },
        "einheit": {
          "type": "string"
        },
        "format": {
          "type": "string",
          "description": "RegEx"
        },
        "bsp": {
          "type": "string",
          "description": "Beispiel"
        }
      }
    },
    "Variable": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "bezeichnung": {
          "type": "string"
        },
        "typ": {
          "$ref": "#/definitions/Typ"
        }
      }
    },
...
  }
}